+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 3.0.37
  - Issue #389 - "Add Missing Parentheses": Skip variables with the same name as a record function. 
+ - Issue #393 - Make FlowFields read-only, remove DataClassification from FlowFields and FlowFilters
+   - new editor and project "Make FlowFields Read-Only..." commands
+   - new "MakeFlowFieldsReadOnly" option for Code Clean-up
+   - DataClassification command removes property from FlowField and FlowFilter table fields
  - Issue #394 - alOutline.codeCleanupActions: tooltips appear to be "off-by-one"
  - Issue #396 - Account for affixes in the variable name-type completion provider(s)
    - "_" character detection in affixes bugfix
@@ -13,6 +17,7 @@
 
 Thank you
  - NKarolak for reporting issue #389
+ - fvet for reporting issue #393
  - jhoek for reporting issue #394
  - rvanbekkum and fvet for reporting issue #396
  - nickgoddard777 for reporting issue #398